Prefers IPv6 addresses, fall back to IPv4.
This commit introduces a new feature flag "IPv6First". When enabled the
`bdns.LookupHost()` function will return addresses with IPv6 first.

When the "IPv6First" feature is enabled the VA's HTTP dialer and TLS SNI
(01 and 02) certificate fetch requests will attempt to automatically
retry when the initial connection was to IPv6 and there is an IPv4
address available to retry with.

The VA and DNS unit tests are updated to test the lookup order change as
well as the fallback connection behaviour.