Tracks addresses tried before the address used in validation records.
This commit adds a `AddressesTried` field to the `ValidationRecord`
struct. The protobuf definitions and serialization code is updated to
reflect the new optional field.

The VA is updated to populate the AddressesTried value when fallback
behaviour is used. The VA fallback unit tests are updated to check for
the correct validation record fields.

A bug whereby the VA wouldn't use the ValidationRecord the dialer
updates with the new AddressUsed is also fixed by adjusting how the
validateChallenge validationRecords are created & returned.